refactor(intelligence): move Acumatica write tools into their own Tools/Acumatica namespace
Per kaioken: group Acumatica-specific write tools separately, mirroring the
existing Tools/FinancialModelingPrep/ pattern, instead of mixing them into
the generic Tools/Accounting/ folder alongside read-only Scribe queries.

Moved (namespace only, no behavior change): CreateApBillTool, VoidApBillTool,
ApplyApPaymentTool, CreateArInvoiceTool, VoidArInvoiceTool, ApplyArPaymentTool.
Tools/Accounting/ now holds only generic, non-ERP-specific read tools
(Find*, List*, Query*, Match*, TopLatePayersTool).
